Move pattern forms to new page
