+++ src/pages/MyPage.vue
@@ -13,13 +13,11 @@
         .row
           .col-12.col-sm-6.col-md-3.q-pa-md.q-my-xs(v-for="(pattern, index) in filteredPatterns" :key="pattern.id")
             PatternCard(:pattern="pattern" :mine="true" @edit="editPattern(pattern._id)")
-    PatternDialog(:open="isModalOpen" :patterndata="editingPattern" @model="val => isModalOpen = val" @refreshPattern="fetchMyPattern")
     q-page-sticky(position="bottom-right" :offset="[36, 36]")
-      q-btn(fab icon="add" color="tech" text-color="black" @click="newPattern")
+      q-btn(fab icon="add" color="tech" text-color="black" @click="$router.push('/patterns/new')")
 </template>
 
 <script>
-import PatternDialog from '../components/PatternDialog'
 import PatternCard from '../components/PatternCard'
 
 export default {
@@ -80,17 +78,14 @@ export default {
     }
   },
   components: {
-    PatternDialog,
     PatternCard
   },
   data () {
     return {
-      isModalOpen: false,
       patterns: [],
       search: '',
       filter: '',
-      error: false,
-      editingPattern: {}
+      error: false
     }
   },
   computed: {
@@ -114,12 +109,7 @@ export default {
       }
     },
     editPattern (id) {
-      this.editingPattern = this.patterns.find(pattern => pattern._id === id)
-      this.isModalOpen = true
-    },
-    newPattern () {
-      this.editingPattern = {}
-      this.isModalOpen = true
+      this.$router.push('/patterns/edit/' + id)
     },
     applySearch () {
       this.filter = this.search
