+++ src/router/routes.js
@@ -5,8 +5,10 @@ const routes = [
     children: [
       { path: '', component: () => import('pages/Index.vue'), meta: { title: 'TECHMANIA', login: false } },
       { path: 'changelog', component: () => import('pages/Changelog.vue'), meta: { title: 'Changelog | TECHMANIA', login: false } },
-      { path: 'patterns', component: () => import('pages/Patterns.vue'), meta: { title: 'Patterns | TECHMANIA', login: false } },
+      { path: 'patterns/new', component: () => import('pages/PatternForm.vue'), meta: { title: 'New Pattern | TECHMANIA', login: true } },
+      { path: 'patterns/edit/:id', component: () => import('pages/PatternForm.vue'), meta: { title: 'Edit Pattern | TECHMANIA', login: true } },
       { path: 'patterns/:id', component: () => import('pages/Pattern.vue'), meta: { title: 'TECHMANIA', login: false } },
+      { path: 'patterns', component: () => import('pages/Patterns.vue'), meta: { title: 'Patterns | TECHMANIA', login: false } },
       { path: 'mypage', component: () => import('pages/MyPage.vue'), meta: { title: 'My Page | TECHMANIA', login: true } }
     ]
   },
